+++ picasso/gui/render.py
@@ -3155,30 +3155,20 @@ class MaskSettingsDialog(QtWidgets.QDialog):
 
     Attributes
     ----------
-    ax1 : plt.axes.Axes
-        Axis where all locs are shown with a given oversampling.
-    ax2 : plt.axes.Axes
-        Axis where blurred locs are shown.
-    ax3 : plt.axes.Axes
-        Axis where binary mask is shown.
-    ax4 : plt.axes.Axes
-        Axis where masked locs are shown (initially shows only zeros).
     cached_blur : int
         0 if image is to be blurred, 1 otherwise.
     cached_oversampling : int
         0 if image is to be redrawn, 1 otherwise.
     cached_thresh : int
         0 if mask is to be calculated, 1 otherwise.
-    canvas : FigureCanvas
-        Canvas used for plotting.
     channel : int
         Channel of localizations that are plotted in the canvas.
     cmap : str
         Colormap used in displaying images, same as in the main window.
+    container : QWidget
+        Container for all widgets in the scroll area.
     disp_px_size : QSpinBox
-        Contains the display pixel size (nm).
-    figure : plt.figure.Figure
-        Figure containing subplots.
+        Contains the display pixel size in nm.
     index_locs : list
         Localizations that were masked; may contain a single or all
         channels.
@@ -3199,7 +3189,7 @@ class MaskSettingsDialog(QtWidgets.QDialog):
     mask : np.array
         Histogram displaying binary mask; displayed in ax3.
     mask_blur : QDoubleSpinBox
-        Contains the blur value.
+        Contains the blur value in nm.
     mask_loaded : bool
         True, if mask was loaded from an external file.
     mask_thresh : QDoubleSpinBox
@@ -3207,6 +3197,11 @@ class MaskSettingsDialog(QtWidgets.QDialog):
     paths : list
         Contains paths to all localizations loaded when starting the
         dialog.
+    pixelsize : float
+        Camera pixel size (nm).
+    plots : list of QLabel
+        Used for displaying the 4 plots (original, blurred, mask,
+        masked).
     save_all : QCheckBox
         If checked, all channels loaded are masked; otherwise only
         one channel.
@@ -3214,8 +3209,6 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         Used for saving masked localizations.
     save_mask_button : QPushButton
         Used for saving the current mask as a .npy file.
-    _size_hint : tuple
-        Determines the recommended size of the dialog.
     window : QMainWindow
         Instance of the main window.
     x_max : float
@@ -3230,34 +3223,40 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         self.setWindowTitle("Generate Mask")
         self.setModal(False)
         self.channel = 0
-        self._size_hint = (670, 840)
-        self.setMinimumSize(*self._size_hint)
 
-        vbox = QtWidgets.QVBoxLayout(self)
-        mask_groupbox = QtWidgets.QGroupBox("Mask Settings")
-        vbox.addWidget(mask_groupbox)
-        mask_grid = QtWidgets.QGridLayout(mask_groupbox)
+        main_layout = QtWidgets.QVBoxLayout(self)
+        scroll = QtWidgets.QScrollArea(self)
+        scroll.setWidgetResizable(True)
+        self.container = QtWidgets.QWidget()
+        scroll.setWidget(self.container)
+        vbox = QtWidgets.QVBoxLayout(self.container)
+        main_layout.addWidget(scroll)
 
-        mask_grid.addWidget(QtWidgets.QLabel("Display pixel size [nm]"), 0, 0)
+        settings_groupbox = QtWidgets.QGroupBox("Settings")
+        vbox.addWidget(settings_groupbox)
+        settings_grid = QtWidgets.QGridLayout(settings_groupbox)
+
+        settings_grid.addWidget(
+            QtWidgets.QLabel("Display pixel size (nm)"), 0, 0,
+        )
         self.disp_px_size = QtWidgets.QSpinBox()
         self.disp_px_size.setRange(10, 99999)
         self.disp_px_size.setValue(300)
         self.disp_px_size.setSingleStep(10)
         self.disp_px_size.setKeyboardTracking(False)
         self.disp_px_size.valueChanged.connect(self.update_plots)
-        mask_grid.addWidget(self.disp_px_size, 0, 1, 1, 2)
+        settings_grid.addWidget(self.disp_px_size, 0, 1, 1, 2)
 
-        mask_grid.addWidget(QtWidgets.QLabel("Blur"), 1, 0)
-        self.mask_blur = QtWidgets.QDoubleSpinBox()
-        self.mask_blur.setRange(0, 9999)
-        self.mask_blur.setValue(1)
-        self.mask_blur.setSingleStep(0.1)
-        self.mask_blur.setDecimals(5)
+        settings_grid.addWidget(QtWidgets.QLabel("Blur (nm)"), 1, 0)
+        self.mask_blur = QtWidgets.QSpinBox()
+        self.mask_blur.setRange(1, 999999)
+        self.mask_blur.setValue(500)
+        self.mask_blur.setSingleStep(10)
         self.mask_blur.setKeyboardTracking(False)
         self.mask_blur.valueChanged.connect(self.update_plots)
-        mask_grid.addWidget(self.mask_blur, 1, 1, 1, 2)
+        settings_grid.addWidget(self.mask_blur, 1, 1, 1, 2)
 
-        mask_grid.addWidget(QtWidgets.QLabel("Threshold"), 2, 0)
+        settings_grid.addWidget(QtWidgets.QLabel("Threshold"), 2, 0)
         self.mask_thresh = QtWidgets.QDoubleSpinBox()
         self.mask_thresh.setRange(0, 1)
         self.mask_thresh.setValue(0.5)
@@ -3265,54 +3264,57 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         self.mask_thresh.setDecimals(5)
         self.mask_thresh.setKeyboardTracking(False)
         self.mask_thresh.valueChanged.connect(self.update_plots)
-        mask_grid.addWidget(self.mask_thresh, 2, 1, 1, 2)
+        settings_grid.addWidget(self.mask_thresh, 2, 1, 1, 2)
 
-        gridspec_dict = {
-            'bottom': 0.05, 'top': 0.95, 'left': 0.05, 'right': 0.95,
-        }
-        (
-            self.figure,
-            ((self.ax1, self.ax2), (self.ax3, self.ax4)),
-        ) = plt.subplots(2, 2, figsize=(6, 6), gridspec_kw=gridspec_dict)
-        self.canvas = FigureCanvas(self.figure)
-        mask_grid.addWidget(self.canvas, 3, 0, 1, 3)
+        display_groupbox = QtWidgets.QGroupBox("Display")
+        vbox.addWidget(display_groupbox)
+        display_layout = QtWidgets.QGridLayout(display_groupbox)
+        self.plots = [QtWidgets.QLabel() for _ in range(4)]
+        display_layout.addWidget(self.plots[0], 0, 0)
+        display_layout.addWidget(self.plots[1], 0, 1)
+        display_layout.addWidget(self.plots[2], 1, 0)
+        display_layout.addWidget(self.plots[3], 1, 1)
 
+        mask_groupbox = QtWidgets.QGroupBox("Mask")
+        vbox.addWidget(mask_groupbox)
+        mask_grid = QtWidgets.QGridLayout(mask_groupbox)
         self.save_all = QtWidgets.QCheckBox("Mask all channels")
         self.save_all.setChecked(False)
-        mask_grid.addWidget(self.save_all, 4, 0)
+        mask_grid.addWidget(self.save_all, 0, 0)
 
         load_mask_button = QtWidgets.QPushButton("Load Mask")
         load_mask_button.setFocusPolicy(QtCore.Qt.NoFocus)
         load_mask_button.clicked.connect(self.load_mask)
-        mask_grid.addWidget(load_mask_button, 5, 0)
+        mask_grid.addWidget(load_mask_button, 1, 0)
 
         self.save_mask_button = QtWidgets.QPushButton("Save Mask")
         self.save_mask_button.setEnabled(False)
         self.save_mask_button.setFocusPolicy(QtCore.Qt.NoFocus)
         self.save_mask_button.clicked.connect(self.save_mask)
-        mask_grid.addWidget(self.save_mask_button, 5, 1)
+        mask_grid.addWidget(self.save_mask_button, 1, 1)
 
         self.save_blur_button = QtWidgets.QPushButton("Save Blurred")
         self.save_blur_button.setEnabled(False)
         self.save_blur_button.setFocusPolicy(QtCore.Qt.NoFocus)
         self.save_blur_button.clicked.connect(self.save_blur)
-        mask_grid.addWidget(self.save_blur_button, 5, 2)
+        mask_grid.addWidget(self.save_blur_button, 1, 2)
 
         mask_button = QtWidgets.QPushButton("Mask")
         mask_button.setFocusPolicy(QtCore.Qt.NoFocus)
         mask_button.clicked.connect(self.mask_locs)
-        mask_grid.addWidget(mask_button, 6, 0)
+        mask_grid.addWidget(mask_button, 2, 0)
 
         self.save_button = QtWidgets.QPushButton("Save localizations")
         self.save_button.setEnabled(False)
         self.save_button.setFocusPolicy(QtCore.Qt.NoFocus)
         self.save_button.clicked.connect(self.save_locs)
-        mask_grid.addWidget(self.save_button, 6, 1, 1, 2)
+        mask_grid.addWidget(self.save_button, 2, 1, 1, 2)
 
         self.cached_oversampling = 0
         self.cached_blur = 0
         self.cached_thresh = 0
         self.mask_loaded = False
+        self.pixelsize = 130
 
     def init_dialog(self) -> None:
         """Initialize dialog when called from the main window. Load
@@ -3321,22 +3323,31 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         self.locs = self.window.view.locs
         self.paths = self.window.view.locs_paths
         self.infos = self.window.view.infos
+        self.pixelsize = self.window.display_settings_dlg.pixelsize.value()
         # which channel to plot
         self.channel = self.window.view.get_channel("Mask image")
         self.cmap = self.window.display_settings_dlg.colormap.currentText()
-        self.show()
         info = self.infos[self.channel][0]
         self.x_max = info["Width"]
         self.y_max = info["Height"]
         self.update_plots()
 
+        # adjust the size of the dialog to fit its contents
+        hint = self.container.sizeHint()
+        self.resize(hint.width() + 45, self.height())
+        # if room is available on the screen, adjust the height as well
+        screen = QtWidgets.QApplication.primaryScreen()
+        screen_height = 1000 if screen is None else screen.size().height()
+        if hint.height() + 45 < screen_height:
+            self.resize(self.width(), hint.height() + 45)
+        else:
+            self.resize(self.width(), screen_height - 100)
+        self.show()
+
     def generate_image(self) -> None:
         """Histogram loaded localizations from a given channel."""
         locs = self.locs[self.channel]
-        oversampling = (
-            self.window.display_settings_dlg.pixelsize.value()
-            / self.disp_px_size.value()
-        )
+        oversampling = self.pixelsize / self.disp_px_size.value()
         viewport = ((0, 0), (self.y_max, self.x_max))
         _, H = render.render(
             locs,
@@ -3345,13 +3356,20 @@ class MaskSettingsDialog(QtWidgets.QDialog):
             blur_method=None,
         )
         self.H = H / H.max()
+        self.plots[0].setPixmap(
+            self.render_to_pixmap(self.H, title="Original")
+        )
 
     def blur_image(self) -> None:
         """Blur localizations using a Gaussian filter."""
-        H_blur = gaussian_filter(self.H, sigma=self.mask_blur.value())
+        blur_px = self.mask_blur.value() / self.disp_px_size.value()
+        H_blur = gaussian_filter(self.H, sigma=blur_px)
         H_blur = H_blur / np.max(H_blur)
-        self.H_blur = H_blur  # image to be displayed in self.ax2
+        self.H_blur = H_blur
         self.save_blur_button.setEnabled(True)
+        self.plots[1].setPixmap(
+            self.render_to_pixmap(self.H_blur, title="Blur")
+        )
 
     def save_mask(self) -> None:
         """Save binary mask to .npy and .png formats."""
@@ -3361,7 +3379,7 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         path, ext = QtWidgets.QFileDialog.getSaveFileName(
             self, "Save mask to", name_mask, filter="*.npy"
         )
-        if path:
+        if path:  # TODO: fix
             np.save(path, self.mask)
             png_path = path.replace(".npy", ".png")
             plt.imsave(png_path, self.mask, cmap="gray")
@@ -3374,7 +3392,7 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         path, ext = QtWidgets.QFileDialog.getSaveFileName(
             self, "Save blur to", name_blur, filter="*.png"
         )
-        if path:
+        if path:  # TODO: fix
             plt.imsave(path, self.H_blur, cmap=self.cmap)
 
     def load_mask(self) -> None:
@@ -3396,13 +3414,13 @@ class MaskSettingsDialog(QtWidgets.QDialog):
             mask[self.H_blur > self.mask_thresh.value()] = 1
             self.mask = mask
             self.save_mask_button.setEnabled(True)
+            self.plots[2].setPixmap(
+                self.render_to_pixmap(self.mask, cmap='Greys_r', title="Mask"),
+            )
 
     def update_plots(self, new_mask: bool = True) -> None:
         """Plot in all 4 axes: 2D histogram, blurred image, mask and
         masked localizations."""
-        if self.mask_blur.value() == 0.00000:
-            self.mask_blur.setValue(0.00001)
-
         if new_mask:
             if self.cached_oversampling:
                 self.cached_oversampling = 0
@@ -3430,21 +3448,6 @@ class MaskSettingsDialog(QtWidgets.QDialog):
             if not self.cached_thresh:
                 self.mask_image()
                 self.cached_thresh = 1
-
-        self.ax1.imshow(self.H, cmap=self.cmap)
-        self.ax1.set_title("Original")
-        self.ax2.imshow(self.H_blur, cmap=self.cmap)
-        self.ax2.set_title("Blurred")
-        self.ax3.imshow(self.mask, cmap='Greys_r')
-        self.ax3.set_title("Mask")
-        self.ax4.imshow(np.zeros_like(self.H), cmap=self.cmap)
-        self.ax4.set_title("Masked image")
-
-        for ax in (self.ax1, self.ax2, self.ax3, self.ax4):
-            ax.grid(False)
-            ax.axis('off')
-
-        self.canvas.draw()
 
     def mask_locs(self) -> None:
         """Mask localizations from a single or all channels."""
@@ -3474,19 +3477,13 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         ):  # update masked locs plot if the current channel is masked
             _, self.H_new = render.render(
                 self.index_locs[-1],
-                oversampling=(
-                    self.window.display_settings_dlg.pixelsize.value()
-                    / self.disp_px_size.value()
-                ),
+                oversampling=self.pixelsize / self.disp_px_size.value(),
                 viewport=((0, 0), (self.y_max, self.x_max)),
                 blur_method=None,
             )
-
-            self.ax4.imshow(self.H_new, cmap=self.cmap)
-            self.ax4.grid(False)
-            self.ax4.axis('off')
-            self.save_button.setEnabled(True)
-            self.canvas.draw()
+            self.plots[3].setPixmap(
+                self.render_to_pixmap(self.H_new, title="Masked")
+            )
 
     def save_locs(self) -> None:
         """Save masked localizations."""
@@ -3570,11 +3567,66 @@ class MaskSettingsDialog(QtWidgets.QDialog):
         info = self.infos[channel] + [{
             "Generated by": f"Picasso v{__version__} Render : Mask {mask_in}",
             "Display pixel size (nm)": mask_pixelsize,
-            "Blur": self.mask_blur.value(),
+            "Blur (nm)": self.mask_blur.value(),
             "Threshold": self.mask_thresh.value(),
             "Area (um^2)": area,
         }]
         return info
+
+    def render_to_pixmap(
+        self,
+        image: np.ndarray,
+        cmap: str = None,
+        title: str = "",
+    ) -> QtGui.QPixmap:
+        """Convert a 2D numpy array to QPixmap for displaying in a
+        QLabel.
+
+        Parameters
+        ----------
+        image : np.ndarray
+            2D array to be converted.
+        cmap : str or None, optional
+            Colormap to be used. If None, self.cmap is used. Default is
+            None.
+        title : str, optional
+            Title of the image, displayed in the top left corner.
+            Default is "".
+
+        Returns
+        -------
+        pixmap : QtGui.QPixmap
+            Converted pixmap.
+        """
+        # adjust contrast and convert to 8 bits
+        if image.dtype != np.int8:
+            image -= image.min()
+            image /= image.max()
+        image = np.round(255 * image).astype("uint8")
+        # get colormap and paint the image
+        cmap = self.cmap if cmap is None else cmap
+        cmap = np.uint8(
+            np.round(255 * plt.get_cmap(cmap)(np.arange(256)))
+        )
+        # create a 4 channel (rgb, alpha) array
+        Y, X = image.shape
+        bgra = np.zeros((Y, X, 4), dtype=np.uint8, order="C")
+        bgra[..., 0] = cmap[:, 2][image]
+        bgra[..., 1] = cmap[:, 1][image]
+        bgra[..., 2] = cmap[:, 0][image]
+        bgra[..., 3] = 255  # set alpha channel to fully opaque
+        qimage = QtGui.QImage(bgra.data, X, Y, QtGui.QImage.Format_RGB32)
+        qimage = qimage.scaled(
+            300, 300, QtCore.Qt.KeepAspectRatioByExpanding,
+        )
+        pixmap = QtGui.QPixmap.fromImage(qimage)
+        if title:
+            painter = QtGui.QPainter(pixmap)
+            painter.setPen(QtGui.QPen(QtCore.Qt.white))
+            painter.setFont(QtGui.QFont('Arial', 15))
+            painter.drawText(10, 20, title)
+            painter.end()
+        return pixmap
 
 
 class PickToolCircleSettings(QtWidgets.QWidget):
@@ -8187,9 +8239,7 @@ class View(QtWidgets.QLabel):
         self._bgra[:, :, 3].fill(255)
         # build QImage
         Y, X = self._bgra.shape[:2]
-        qimage = QtGui.QImage(
-            self._bgra.data, X, Y, QtGui.QImage.Format_RGB32
-        )
+        qimage = QtGui.QImage(self._bgra.data, X, Y, QtGui.QImage.Format_RGB32)
         return qimage
 
     def read_colors(self, n_channels: int | None = None) -> list[list[float]]: